+++ src/types/messages.ts
@@ -6,11 +6,11 @@ import type {
   MsgWithdrawValidatorCommission,
   MsgFundCommunityPool,
 } from '../proto/cosmos/distribution/v1beta1/tx_pb';
-import { MsgVote, MsgVoteWeighted, MsgDeposit } from '../proto/cosmos/gov/v1beta1/tx_pb';
+import { MsgVote, MsgVoteWeighted, MsgDeposit, MsgSubmitProposal } from '../proto/cosmos/gov/v1beta1/tx_pb';
 import { MsgUnjail } from '../proto/cosmos/slashing/v1beta1/tx_pb';
 import { MsgEditValidator, MsgDelegate, MsgBeginRedelegate, MsgUndelegate } from '../proto/cosmos/staking/v1beta1/tx_pb';
 import { MsgCreateVestingAccount } from '../proto/cosmos/vesting/v1beta1/tx_pb';
-import { Proposal } from '../proto/cosmos/gov/v1beta1/gov_pb';
+import { Proposal, TextProposal } from '../proto/cosmos/gov/v1beta1/gov_pb';
 import { MsgGrant } from '../proto/cosmos/authz/v1beta1/tx_pb';
 import { Coin } from '../proto/cosmos/base/v1beta1/coin_pb';
 
@@ -20,6 +20,8 @@ export type MsgSetWithdrawAddressDisplay = MsgSetWithdrawAddress.AsObject;
 export type MsgWithdrawDelegatorRewardDisplay = MsgWithdrawDelegatorReward.AsObject;
 export type MsgWithdrawValidatorCommissionDisplay = MsgWithdrawValidatorCommission.AsObject;
 export type MsgFundCommunityPoolDisplay = MsgFundCommunityPool.AsObject;
+export type MsgSubmitProposalDisplay = MsgSubmitProposal.AsObject & { proposalType: 'TextProposal' | undefined };
+export type TextProposalDisplay = TextProposal.AsObject;
 export type ProposalDisplay = Proposal.AsObject;
 export type MsgVoteDisplay = MsgVote.AsObject;
 export type MsgVoteWeightedDisplay = MsgVoteWeighted.AsObject;