+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add MsgSubmitProposal to message-service [#43] (https://github.com/provenance-io/wallet-lib/issues/45)
+
 ## 2.1.0
 
 - Added weighted votes to message-service [#42] (https://github.com/provenance-io/wallet-lib/issues/42)